requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
pandas
python-dotenv
chainlit
literalai
nltk==3.8.1
PyPDF2
pypdf
fuzzywuzzy[speedup]
duckduckgo-search
ragas
guardrails-ai


